refactored contacts code
