+++ thread/src/components/Common/ContentSpinner.js
@@ -1,0 +1,11 @@
+import React from 'react';
+import { Spinner } from 'native-base';
+import CentredContent from './CentredContent';
+
+const ContentSpinner = ({ color }) => (
+	<CentredContent>
+        <Spinner color={color} />
+    </CentredContent>
+);
+
+export default ContentSpinner;
