+++ thread/src/components/Common/index.js
@@ -1,0 +1,3 @@
+export { default as CommonButton } from './CommonButton';
+export { default as CentredContent } from './CentredContent';
+export { default as ContentSpinner } from './ContentSpinner';
