+++ thread/src/components/Contacts/RequestsReceived.js
@@ -1,39 +1,27 @@
 import React, { Component } from 'react';
-import { Container, Content, List, ListItem, Left, Body, Right, Thumbnail, Text, H3, Button, Spinner } from 'native-base';
 import { connect } from 'react-redux';
+import { Container, Content, List, ListItem, Body, Right, Text, Button } from 'native-base';
+import { CentredContent, ContentSpinner } from '../Common';
 import {
     acceptRequest
-} from '../actions';
+} from '../../actions';
 
 class RequestsSent extends Component {
+
     onAcceptPress = (requestorID) => {
         acceptRequest(requestorID, this.props.chatUser);
     }
 
     renderList() {
         if (this.props.loading) {
-            return (
-                <Content contentContainerStyle={{ 
-                    flex: 1,
-                    flexDirection: 'column',
-                    justifyContent: 'center',
-                    alignItems: 'center'
-                }}>
-                    <Spinner color="blue" />
-                </Content>
-            );
+            return <ContentSpinner color="blue" />;
         }
 
         if (this.props.requests_received_list.length === 0) {
             return (
-                <Content contentContainerStyle={{ 
-                    flex: 1,
-                    flexDirection: 'column',
-                    justifyContent: 'center',
-                    alignItems: 'center'
-                }}>
+                <CentredContent>
                     <Text>No requests received</Text>
-                </Content>
+                </CentredContent>
             );
         }
         
@@ -65,7 +53,7 @@ class RequestsSent extends Component {
 
     render() {
         return (
-            <Container style={{ backgroundColor: '#fff' }}>
+            <Container>
                 {this.renderList()}
             </Container>
         );
