+++ thread/src/components/common/CommonButton.js
@@ -2,9 +2,9 @@ import React from 'react';
 import { Button, Icon } from 'native-base';
 
 const CommonButton = ({onPress, name}) => (
-  <Button transparent onPress={onPress}>
-    <Icon name={name} />
-  </Button>
+	<Button transparent onPress={onPress}>
+		<Icon name={name} />
+	</Button>
 );
 
-export default CommonButton
+export default CommonButton;
